requirements.txt
# =============================
# 🔌 Neo4j 图数据库连接驱动
# =============================
neo4j==5.13.0          # 官方推荐的Neo4j Python驱动
py2neo==2021.2.4

# =============================
# 📄 PDF 文本处理
# =============================
pdfplumber==0.9.0       # 推荐使用，基于pdfminer，功能更完善
pdfminer.six==20221105  # pdfplumber 的底层依赖，保留

# =============================
# ✨ 信息抽取与自然语言处理
# =============================
spacy==3.5.0
nltk==3.8.1
transformers==4.30.2
torch==2.0.1            # CPU/GPU 通用，可根据环境替换为cu版本

# 中文分词支持（模型需手动下载）
# ▶ 运行后需执行：python -m spacy download zh_core_web_sm

# 英文分词模型
# ▶ 运行后需执行：python -m spacy download en_core_web_sm

# =============================
# 🧠 文本预处理与增强工具
# =============================
tqdm==4.65.0
python-dotenv==1.0.0
unidecode==1.3.6
pandas==2.0.3

# =============================
# 🔍 实体模糊匹配工具（用于实体对齐、标准化）
# =============================
fuzzywuzzy==0.18.0
python-Levenshtein==0.20.9

# =============================
# 🧠 语义理解（可选）
# =============================
sentence-transformers==2.2.2  # 如果需要向量搜索、语义匹配
openai==0.27.8                # 如需接入 ChatGPT / GPT API 可保留

# =============================
# 🖼️ Web 可视化界面
# =============================
streamlit==1.22.0
pyvis==0.3.2
